put log into module; fixed migration
